refactor: event 도메인 폴더 위치 변경
