+++ src/main/java/com/scg/stop/event/repository/EventPeriodRepository.java
@@ -1,6 +1,6 @@
-package com.scg.stop.domain.event.repository;
+package com.scg.stop.event.repository;
 
-import com.scg.stop.domain.event.domain.EventPeriod;
+import com.scg.stop.event.domain.EventPeriod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventPeriodRepository extends JpaRepository<EventPeriod, Long> {
